create login and adjust relations

+++ src/usuario/usuario.service.ts
@@ -3,9 +3,11 @@ import {
   Inject,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreateUsuarioDto } from './dto/create-usuario.dto';
 import { FindByEmailDto } from './dto/find-by-email.dto';
+import { LoginDto } from './dto/login.dto';
 import { updatePasswordDto } from './dto/update-password.dto';
 import { UpdateUsuarioDto } from './dto/update-usuario.dto';
 import { UsuarioEntity } from './entities/usuario.entity';
@@ -25,13 +27,21 @@ export class UsuarioService {
     if (findUser) {
       throw new ConflictException('Este usuário já existe na base de dados!');
     }
-    const crypto = require('crypto');
-    const sha256Hash = crypto.createHash('sha256');
-    const hashPassword = sha256Hash
-      .update(createUsuarioDto.password)
-      .digest('hex');
+    const hashPassword = await this.hasher(createUsuarioDto.password);
     createUsuarioDto.password = hashPassword;
     return await this.usuarioRepoService.create(createUsuarioDto);
+  }
+
+  async login(loginDto: LoginDto): Promise<boolean> {
+    let findUser = await this.usuarioRepoService.findByEmail(loginDto.email);
+    if (!findUser) {
+      throw new UnauthorizedException('Usuário ou Senha Incorretos!');
+    }
+    const loginPassword = await this.hasher(loginDto.password);
+    if (loginPassword !== findUser.password) {
+      throw new UnauthorizedException('Usuário ou senha Incorretos!');
+    }
+    return true;
   }
 
   async findAll(): Promise<UsuarioEntity[]> {
@@ -76,6 +86,19 @@ export class UsuarioService {
     if (findUser) {
       throw new NotFoundException('Usuário não encontrado!');
     }
+
+    if (findUser.clientes.length > 0) {
+      throw new UnauthorizedException(
+        'Não é possível excluir usuário com propostas abertas!',
+      );
+    }
     return await this.usuarioRepoService.delete(id);
+  }
+
+  hasher(text: string): Promise<string> {
+    const crypto = require('crypto');
+    const sha256Hash = crypto.createHash('sha256');
+    const hashPassword = sha256Hash.update(text).digest('hex');
+    return hashPassword;
   }
 }
